Seperated Piece logic and helpers
